refactor: challenges 로 패키지명 변경
