+++ src/main/java/WebProject/withpet/challenges/dto/WeeklyChallengeResponseDto.java
@@ -1,6 +1,6 @@
-package WebProject.withpet.challenge.dto;
+package WebProject.withpet.challenges.dto;
 
-import WebProject.withpet.challenge.domain.Challenge;
+import WebProject.withpet.challenges.domain.Challenge;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -8,25 +8,23 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 @Builder
-public class DailyChallengeResponseDto {
+public class WeeklyChallengeResponseDto {
     private Long challengeId;
-    private Long challengeLogId;
+
     private Long petId;
     private String title;
     private int targetCnt;
     private int achieveCnt;
-    private boolean isAchieved;
 
-    public DailyChallengeResponseDto(
+    public WeeklyChallengeResponseDto(
             Challenge challenge,
-            int achieveCnt, Long challengeLogId, Long petId
+            int achieveCnt,
+            Long petId
     ) {
         this.challengeId = challenge.getId();
         this.petId = petId;
         this.title = challenge.getTitle();
         this.targetCnt = challenge.getTargetCnt();
         this.achieveCnt = achieveCnt;
-        this.challengeLogId = challengeLogId;
-        this.isAchieved = challengeLogId != null;
     }
 }
